pyproject.toml
[tool.poetry]
name = "zuno"
version = "2.2.0"
description = "A personal agent workspace platform for MCP, skills, knowledge and desktop execution"
authors = ["ProfessorZhi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^1.0.3"
langgraph = "1.0.10"
langgraph-prebuilt = "1.0.8"
langchain-community = "^0.4.1"
pdf2docx = "^0.5.8"
loguru = "^0.7.3"
langchain-openai = "^1.0.2"
dashscope = "^1.25.0"
tavily-python = "^0.7.12"
mcp = "^1.20.0"
fastapi = "^0.121.0"
sqlmodel = "^0.0.27"
alembic = "^1.16.4"
pytz = "^2025.2"
redis = "^7.0.1"
aio-pika = "^9.5.5"
asyncpg = "^0.30.0"
oss2 = "^2.19.1"
pypandoc = "^1.15"
elasticsearch = "^9.2.0"
pymilvus = "^2.6.3"
pymupdf4llm = "^0.1.8"
aiofiles = "^25.1.0"
chromadb = "^1.3.4"
requests-html = "^0.10.0"
playwright = "^1.58.0"
selenium = "^4.38.0"
pymysql = "^1.1.2"
aiomysql = "^0.3.2"
psycopg = {version = "^3.2.9", extras = ["binary"]}
google-search-results = "^2.4.2"
arxiv = "^2.3.0"
anthropic = "^0.72.0"
langsmith = "^0.4.8"
neo4j = "^5.28.1"
html2image = "^2.0.7"
lxml-html-clean = "^0.4.3"
pyfiglet = "^1.0.4"
minio = "^7.2.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
